doc(schema): Export canonical event schema as json schema
This introduces schemars as dependency and adds a new subcommand to
`relay` that generates the schema, which we then use to publish the
schema as release artifact.

Alex Hofsteede attempted to document the event schema using JSON schema
too, but sort of got stuck halfway through. I believe the difference is
that we only attempt to document the canonical version *after* store
normalization. The only exception we make there is that
tags/headers/cookies are also allowed as objects even though the
canonical form is array of arrays.

Every attribute has to be decorated for both our own derive and
schemars' derives. For example, for each `metastructure(field = foo)`
one now also has to write `schemars(rename = foo)`. Also every field now
has to be decorated with `schemars(default)` to allow for missing
values. While I think it would be nice if e.g. field renames were synced
across both systems, I believe there may be a few cases where we
intentionally want to make a field hard-required in the JSON schema
without actually enforcing it on the server.

This schema file's target audience is mostly SDK developers, but maybe
Snuba ppl can profit from it too.

Note: We cannot profit from this within the Python SDK as the only
tooling for using JSON schema in type annotations is
https://pypi.org/project/jsonschema-typed/ which users would have to
install in their devenv to see any effect. What I actually wanted was
code generation from JSON schema, not a runtime plugin for the type
checker.